Move unshift tests to a separate file
